# saas-multi-tenant-app/src/main/java/com/rks/saas_multi_tenant_app/requests/CategoryRequest.java
package com.rks.saas_multi_tenant_app.requests;


import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class CategoryRequest {

    @NotBlank(message = "Category name should not be empty")
    @Size(min = 3, max = 255, message = "Category name should be between 3 and 255 characters")
    private String name;
    private String description;
}