# saas-multi-tenant-app/src/main/java/com/rks/saas_multi_tenant_app/responses/ProductResponse.java
package com.rks.saas_multi_tenant_app.responses;


import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.math.BigDecimal;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class ProductResponse {

    private String id;
    private String name;
    private String reference;
    private String description;
    private Integer alertThreshold;
    private BigDecimal price;
    private String categoryId;
    private String categoryName;
    private int availableQuantity;
}